+++ app/(screens)/location/[location].tsx
@@ -3,7 +3,7 @@ import { useGlobalSearchParams, useRouter } from "expo-router";
 import { deleteLocationById } from "@/app/database/locationService";
 import { useState } from "react";
 import LocationView from "@/app/components/location/location-view";
-import { View } from "react-native";
+import { Alert, View } from "react-native";
 
 export default function LocationDetails() {
   const router = useRouter();
@@ -16,13 +16,29 @@ export default function LocationDetails() {
   const deleteLocationClicked = async () => {
     if (!locationObj) return;
 
-    try {
-      await deleteLocationById(locationObj.id);
-      console.log("Location deleted");
-      router.replace("/(tabs)/locations");
-    } catch (error) {
-      console.error("Failed to delete location:", error);
-    }
+    Alert.alert(
+      "Confirm Deletion",
+      "Are you sure you want to delete this location?",
+      [
+        {
+          text: "Cancel",
+          style: "cancel",
+        },
+        {
+          text: "Delete",
+          onPress: async () => {
+            try {
+              await deleteLocationById(locationObj.id);
+              console.log("Location deleted");
+              router.replace("/(tabs)/locations");
+            } catch (error) {
+              console.error("Failed to delete location:", error);
+            }
+          },
+        },
+      ],
+      { cancelable: true }
+    );
   };
 
   return (
